+++ NEWS.md
@@ -1,4 +1,5 @@
 # teal 0.9.5.9000
+* Replaced `random.cdisc.data` with `scda` in examples and tests.
 
 ### New features
 * Added `cdse_dataset_connector` to create delayed data objects from `CDSE`.